Release Safari 10.0 (#1005)
* Fix error while running npm commands witch create local node_modules

* Bump wtm packages

* Bump xcode build version

* Add vite patch to get glob imports after static ones

* Add async attribute for faster in order execution

* Bump xcode build version

+++ tools/licenses.js
@@ -72,7 +72,9 @@ const template = (packages) => `
 </html>
 `;
 
+let symlink = false;
 if (!fs.existsSync('./node_modules') && fs.existsSync('../node_modules')) {
+  symlink = true;
   fs.symlinkSync('../node_modules', './node_modules', 'dir');
 }
 
@@ -108,8 +110,9 @@ checker.init(
     }
 
     fs.writeFileSync('./dist/licenses.html', template(output));
-    fs.unlinkSync('./node_modules');
-
+    if (symlink) {
+      fs.unlinkSync('./node_modules');
+    }
     console.log('Completed licenses.html');
   },
 );
